+++ src/main/java/com/workable_sb/workable/repositories/GeneroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.Genero;
 
-public interface GeneroRepository extends JpaRepository <Genero, Long> {
+public interface GeneroRepository extends JpaRepository <Genero, Integer> {
 
 }
