+++ src/main/java/com/workable_sb/workable/repositories/InfoPersonalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.InfoPersonal;
 
-public interface InfoPersonalRepository extends JpaRepository<InfoPersonal, Short> {
+public interface InfoPersonalRepository extends JpaRepository<InfoPersonal, Integer> {
 
 }
